# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get a user by its id
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: User id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid id supplied
        "401":
          description: User not authorized
        "403":
          description: Access denied
        "404":
          description: User not found
    put:
      tags:
      - user-rest-controller
      summary: Update a user by its id
      operationId: replaceUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "200":
          description: User updated correctly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: User not updated
        "403":
          description: User not authorized
    delete:
      tags:
      - user-rest-controller
      summary: Deletes a user by its id
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted correctly
        "400":
          description: Bad request
        "401":
          description: User not authorized
        "403":
          description: User not authorized
        "404":
          description: User not found
...
  /api/users/new:
    post:
      tags:
      - user-rest-controller
      summary: Registers a new user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDTO"
        required: true
      responses:
        "201":
          description: User registered correctly
        "400":
          description: "Bad request, maybe one of the user attributes is missing or\
            \ the type is not valid"
        "409":
          description: User already exists
...
  /api/auth/refresh:
    post:
      tags:
      - login-controller
      operationId: refreshToken
      parameters:
      - name: RefreshToken
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
          readOnly: true
        email:
          type: string
          description: Email address of the user
          example: john.doe@example.com
        password:
          type: string
          description: Password for authentication (write-only field)
          writeOnly: true
        weight:
          type: number
          format: double
        goalWeight:
          type: number
          format: double
        height:
          type: integer
          format: int32
        gender:
          type: string
        age:
          type: integer
          format: int32
        morphology:
          type: string
        caloricPhase:
          type: string
      required:
      - email
      - password
    Food:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        calories:
          type: integer
          format: int32
        type:
          type: string
        listNutritions:
          type: array
          items:
            $ref: "#/components/schemas/Nutrition"
    Nutrition:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 50
          minLength: 3
        type:
          type: string
        user:
          $ref: "#/components/schemas/User"
        listFoods:
          type: array
          items:
            $ref: "#/components/schemas/Food"
      required:
      - name
      - type
    Training:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 50
          minLength: 3
        intensity:
          type: string
          pattern: ^(Baja|Moderada|Alta)$
        duration:
          type: integer
          format: int32
          maximum: 300
          minimum: 1
        exercises:
          type: string
          maxLength: 500
          minLength: 0
        goal:
          type: string
          pattern: ^(Bajar de peso|Mantenerse|Subir de peso)$
        user:
          $ref: "#/components/schemas/User"
      required:
      - exercises
      - goal
      - intensity
      - name
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imgUser:
          type: object
          properties:
            binaryStream:
              type: object
        name:
          type: string
          maxLength: 50
          minLength: 3
        email:
          type: string
        weight:
          type: number
          format: double
          minimum: 1
        goalWeight:
          type: number
          format: double
          minimum: 1
        height:
          type: integer
          format: int32
          maximum: 300
          minimum: 1
        gender:
          type: string
        age:
          type: integer
          format: int32
          maximum: 150
          minimum: 1
        morphology:
          type: string
        caloricPhase:
          type: string
        nutritionList:
          type: array
          items:
            $ref: "#/components/schemas/Nutrition"
        trainingList:
          type: array
          items:
            $ref: "#/components/schemas/Training"
        roles:
          type: array
          items:
            type: string
      required:
      - age
      - caloricPhase
      - gender
      - goalWeight
      - height
      - morphology
      - name
      - weight
...
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...